chore: model changes such that versions are now stored as long instead of as int

+++ backend/src/main/java/com/boardwise/backend/vault/dto/response/AcquireWriteLockDto.java
@@ -11,5 +11,5 @@ public class AcquireWriteLockDto {
     private boolean lockGranted;
     private String lockedBy; // Username instead of id
     private Instant expiresAt;
-    private int currentVersion;
+    private long currentVersion;
 }
